linguis/pyproject.toml
[project]
name = "linguis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ted Neward", email = "ted@tedneward.com" }
]
requires-python = ">=3.13"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
]

[project.scripts]
linguis = "linguis:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
